+++ cypress/e2e/error.cy.js
@@ -1,11 +1,14 @@
 describe('Error Page Test', () => {
-  it('check error routing', () => {
+  beforeEach(() => {
     cy.visit('/wrong-url');
-    cy.url().should('include', '/404');
-    cy.get('h1').contains('Error Page');
   });
 
-  it('Does not do much!', () => {
-    expect(true).to.equal(true);
+  afterEach(() => {
+    cy.window().trigger('unload');
+  });
+
+  it('check error routing', () => {
+    cy.url().should('include', '/404');
+    cy.get('h1').contains('Error Page');
   });
 });
