+++ src/components/card-expanded/card-expanded.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
 
-import { Product } from '../../types/data';
+import { Product } from '@/types/data';
 import { CardExpanded } from './card-expanded';
 
 const cardItem: Product = {
