+++ src/components/card/card.test.tsx
@@ -1,7 +1,8 @@
 import { describe, it, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
+
 import { Card } from './card';
-import { Product } from '../../types/data';
+import { Product } from '@/types/data';
 
 const cardItem: Product = {
   id: 11,
