+++ src/components/card/card.tsx
@@ -1,4 +1,4 @@
-import { Product } from '../../types/data';
+import { Product } from '@/types/data';
 
 interface Props {
   handleOpen: (id: number) => void;
