+++ src/components/form/form.test.tsx
@@ -1,7 +1,8 @@
 import { describe, it, vi } from 'vitest';
 import { screen } from '@testing-library/react';
-import { Form } from './form';
 import { renderWithProviders } from '@/utils/test-utils';
+
+import { Form } from './form';
 
 describe('Form', () => {
   window.URL.createObjectURL = vi.fn();
