+++ src/components/header/header.test.tsx
@@ -1,7 +1,8 @@
 import { describe, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { Header } from './header';
 import { MemoryRouter } from 'react-router-dom';
+
+import { Header } from './header';
 
 describe('Header', () => {
   it('Renders header with diaplyed title', async () => {
