+++ src/components/icons/index.ts
@@ -1,2 +1,3 @@
+/// <reference types="vite-plugin-svgr/client" />
 import { ReactComponent as Spinner } from '@/assets/double-ring.svg';
 export { Spinner };
