+++ src/components/toast/toast.test.tsx
@@ -1,12 +1,11 @@
-import React from 'react';
-
 import { describe, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
+
 import { Toast } from './toast';
 
 describe('Toast component', () => {
   it('Render toast', () => {
-    render(<Toast showToast={true} />);
+    render(<Toast />);
     expect(screen.getByText(/card succesfully added/i)).toBeInTheDocument();
   });
 });
