+++ src/pages/error/errorPage.test.tsx
@@ -1,5 +1,6 @@
 import { describe, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
+
 import { ErrorPage } from './errorPage';
 
 describe('Error page', () => {
